Split handle and handler into two separate things.
Term now has a concept of a Handler literal containing a set of handle
clauses.
Handle now takes an arbitrary term that must evaluate to a handler.

This opens the door for handler reuse (they can be stored in a
variable), and possibly replacing explicit handler literals with plain
records that conform to the interface.

It is motivated by easing the elaboration of effects into core because
we can now elaborate a handler without having to figure out how to
elaborate a `handle ... with ...`.