Introduce new primitives in Core for effects.
Alongside these changes:
  * A small bug fix for row evidence (it wasn't unwrapping singleton
    products).
  * We emit some code for handlers and ops now, but it doesn't work yet
      - Still have to figure out the evidence tracking stuff.

NewPrompt - generates a unique prompt
Prompt - intalls a prompt in execution context and evaluated it's body
under it
Yield - yields to the prompt installed for specified marker capturing
continuation up to that marker.

Eventually we won't need these as Core primitives, but this provides a
conceptual stop gap during elaboration. Emitting these is easier then
proceeding with the monadic translation (mostly because I do not
understand how to do the monaidc translation yet).