Making more progress on effects.
The generated core is now frustratingly close to what we want (with the
side benefit of fixing a lot of bugs in the typechecker). However it's
unclear to me how to resolve the last couple issues.

I've started work on a simple direct interpreter with the hopes that
viewing some execution traces will enlighten me as to how to fix the
final issues.

Most notably, how do we handle "return" in clauses, atm it's unused
which leads to us returning an unbound type variable when our body
should result in IntTy.