BIG CHANGE: support module system in nameres
No syntax or filesystem stuff for modules yet, but this commit adds data
structures for the module tree and module contents, and logic for
resolving names against the module system. Needs some testing once we
have syntax and systems for creating and importing modules.

Also refactors nameres to use opaque IDs instead of pointer handles.
We'll probably use handles again though for string interning, so not
deleting that module.

Also splits the CST into a few new types for resolved versions, since
after nameres we can determine, e.g., whether an identifier refers to a
variable or a top-level item, or whether a dot-access is a module-member
access or not.