Add support for primitives (+) and (<). (<) returns a "builtin" boolean type (which is just an alias for <True |> {}, False |> {}>). Tests show this is working as intended.
We don't need prims, but they've been added to allow for more complex
logic to be tested which will be handy as we want to test multiple
effects working together. For example, with the prims support we can
write an iterator effect that yields an increasing int until it gets above a
threshold.