Code generation for exampleMVPMultiEffect is working.
The issue was during code gen of a `Handle` node we didn't connect the
external evv to the internal evv correctly when updating evv.

Previously we would generate a system of constraints:
```
  external ~ a :⊙ "State"
  internal ~ a :⊙ "State"
```
with the idea to use the external constraint as evidence to remove any
existing "State" handler before installing our new "State" handler,
using the internal constraint as evidence.

This fails because we don't have enough information to figure out what
`a` should be so our system fails to produce concrete evidence. The
solution is to instead construct one piece of evidence of the form
```
  external ~ internal :⊙ "State"
```
This system is solvable and produces the concrete evidence for evv
operations we expect.

TODO: This will break if evv already has a "State" in it with our
current `:⊙` behavior. Look into how to fix this, preferably without
introducing a new kind of constraint.