Big architecture change to type checking.
Previously we used an immutable oriented bag of constraints that gets
solved to a substitution. This works well in haskell, but has O(N^2)
runtime for degenerate cases which is bad.

New type checker uses a union-find datastructure to do unificaiton in
near linear time with mutation. This will be much faster and more
idiomatic moving forward.

Code still needs some clean up, but this lays out the big idea and is
confirmed working on simple test cases.