Initial support for ProdTys, this is rough around the edges and not finished yet.
The big addition in this commit is row unification.
This involves adding a new constraint `RowCombine` that represents two
rows should combine to equal a third.
Because of this we also expand our unification table to store the new
`UnifyVal` in place of just a `Ty`. A `UnifyVal` is either a Ty or a
PartialRow which represents the row a unifier variable will solve into.

More tests are needed for the row polymorphic stuff but that will
require expanding our type checker to produce Schemes (so that they can
contain constraints).