More robust row type support.
We can now type check the classic wand example `(m * n).x` which was not
possible prior with prior row types. This required a substantial change
to unificaiton. Previously a unifier variable stood for either a type or
a row. However this does not accurately represent rows as multiple
unifications of a row can be true (to support all the ways a row can be
created and projected). To support the multiplicity of rows, a unifier
now stands for either a type or a ROW SET. A row set is the set of all
predicates that have to be true about a row.