Track each type of constraint separately.
Constraints now has members `tys` and `rows`, each collect type
equalities and row combination respectively. This allows our solving to
unify all type equalities and then all row combinations.

This solves an issue we have with propagating a row solution when a type
equality makes one of it's variables a concrete row after it's been
added to the unification table.