Construct IR row evidence terms for solved rows.
During lowering we walk our term and collect all the row evidence
used by the term. Then check which of them are solved (and therefore
won't be parameters in the signature). For solved row evidence we
construct a 4-tuple of combinators that provide the functionality we'll
need for any possible row operation on that set of rows. For example if
we have evidence `Row { left: a, right: b, goal: c }` we produce:

```
( concat: {a} -> {b} -> {c}
, branch: (<a> -> t) -> (<b> -> t) -> (<c> -> t)
, ( prj_l: {c} -> {a}
  , inj_l: <a> -> <c>
  )
, ( prj_r: {c} -> {b}
  , inj_r: <b> -> <c>
  )
)
```

Each term (Concat, Branch, etc.) desconstructs this tuple to get the
particular combinator it needs and applies it's arguments to it, producing the
output we need.