Initial pass at lowering effect AST nodes into delim cont IR.
Doing this intial pass at lowering effects has revealed some issues with
how we do type checking. We need to communicate the return variable for
a handler to the operations that will yield to that handler. That allows
us to properly type the continuation that gets captured by it's answer
type.

I think the resolution for this isn't bad tho. We currently don't track
any info in the effect row. If we instead track the return type for the
handler in the effect row, then any operation can find the return type of it's innermost handler by looking at whatever the effect row says.

This while require some more work in the row unification though. As
always the answer to our problems it make our row types more robust.