Fix lower tests, and improve `ir_matcher!`.
`lower_wand` was failing because we did not have a stable order for
constraints in schemes causing the parameter order to change each time
we ran the test. Scheme constraints are now ordered so lowered parameter
list is in a stable order.

ir_matcher! now makes it easier to check varaibles are equal. If a
pattern binds a variable and then later refers to that same ID,
ir_matcher will rebind that variable to a unique pattern and check that
it's equal to the binding. This removes a lot of superfluous assert_eqs
in match bodies.