Change row unification architecture.
Previously we stored rows, and row solutions in the unificaiton table.
This turns out to be a bad fit. We need to remove solved equations from
the unification table as we learn enough about them to decompose them
into trivial `var ~ row` equalities. Which is not idiomatic to represent
with our unification table.

Now we store unsolved equations in a separate structure and anytime we
learn new facts about rows we consult the unsolved equations and solve
any of them that we need to, removing them from the set.

This is not only much easier to represent (look at how much code we can
delete), it also handles back propagation cases that we couldn't handle
with the previous system.