Add new `Handles` constraint.
This is a constraint in the type checker that a type handles an effect.
It is used when type checking a `Handle` term to allow us to remember
what handlers map to what effect variables and wait till constraint
solving to figure out what type our handler has.

This allows us to use variables as handlers as we don't have to have a
concrete type during constraint generation of the handler.