Associate a span with each constraint we generate & desugar changes.
We now associate a span with each constraint we generate for error
tracking purposes. This will allow us to produce diagnostics when we
encounter type checking errors during unification.

Included in this commit is a substantial change to desugar to ensure
each ast node we produce is associated with a span. This often leads to
multiple ast nodes mapping to the same span as they are produced from a
single nst node and there is not an obvious way to subdivide the span.