Introduce salsa and core salsa structs.
This adds salsa-2022 as a depedency of aiahr_core.
It introduces two salsa structs in core, `Ident` and `Module`. `Ident`
is an interned identifier. `Module` stores metadata we need as input for
a module and acts to tie together metadata from each pass for a module.

This should unblock migrating each pass independently as they can use
these structs at API boundaries and migrate their internals.