Parameterize type structs by their allocation strategy.
To store our types in Salsa we need them to have no lifetimes and no
generics (at time of writing). However while doing inference we'd like
types to be parameterized by the lifetime of the inference context so we
can be confident internals of inference don't leak outside type
checking. This is helpful to ensure unification variables are properly
handled by type checking, etc.

To accomplish both these goals we introduce a TypeAlloc trait with
associated types for the low level storage of type data and row data.
Things like the type of closed row's fields and values, or the data that
`Ty` wraps.

This trait has two implementations, `InArena<'_>` and `InDb`. As their
names imply, `InArena<'ctx>` stores everything in an arena with lifetime
`'ctx`. Whereas InDb, stores things as tracked structs in salsa. This
strategy allows us to still move between allocations via `TypeFoldable`
and the output of type checking is now a suite of `InDb` types that
can be tracked by salsa without lifetime parameters.

This change still needs some cleaning up, it's a pretty minimal set of
changes to proof of concept the idea.