Swap `parser` over to use `Ident` (#22)
* Add `salsa` as dependency of `parser`.

* Swap `parser` over to use `Ident`.

This completes our swap to `Ident`, we not produce them in the lexer and
use them throughout each pass. Because of this changes propagate outside
parser quite a bit. We clean up all our conversion points and by
extension are able to remove some references to db and interner.

* Fix cargo clippy warnings.

* Parameterize type structs by their allocation strategy.

To store our types in Salsa we need them to have no lifetimes and no
generics (at time of writing). However while doing inference we'd like
types to be parameterized by the lifetime of the inference context so we
can be confident internals of inference don't leak outside type
checking. This is helpful to ensure unification variables are properly
handled by type checking, etc.

To accomplish both these goals we introduce a TypeAlloc trait with
associated types for the low level storage of type data and row data.
Things like the type of closed row's fields and values, or the data that
`Ty` wraps.

This trait has two implementations, `InArena<'_>` and `InDb`. As their
names imply, `InArena<'ctx>` stores everything in an arena with lifetime
`'ctx`. Whereas InDb, stores things as tracked structs in salsa. This
strategy allows us to still move between allocations via `TypeFoldable`
and the output of type checking is now a suite of `InDb` types that
can be tracked by salsa without lifetime parameters.

This change still needs some cleaning up, it's a pretty minimal set of
changes to proof of concept the idea.

* Clean up tc code.

* Remove unneeded 's lifetimes

* Split tc out into more submodules.

* Split lower out into submodules.

* Move lower module into it's own file

* Dang it.