Move bulk of type structs into core.
This move is a little awkward. We have two variants of type structs,
InArena and InDb. We'd like InArena to be private within aiahr_tc.
However this doesn't work out, the functionality we want on arena types
must be defined in the crates they come from.

To get around this, the InArena types are feature gated and that feature
is only enabled when core is imported in aiahr-tc. This isn't an ideal
solution but achieves some encapsulation.

NB: aiahr-ty didn't work, it's gone.