Create AST Item enum.
This makes the parallels between nst and ast more apparent. `Ast` acts
as the Ast parallel of `nst::Item::Term`. This means the nst annotation
and name will be desugared into members of the `Ast` struct.

This commit also adds scheme annotations to AST defaulting to None since
they are optionally. This will be used when we desugar a syntactic
type into a semantic type in desugar.