Replace NstModule with NameResModule.
This change allows us to store name res types in our NameResModule but
anything that wants to use them now has to take a dependency on
aiahr-analysis.

Once NameResModule is more stable and we're confident in what we want to
store their we can move it into core to remedy this.