Create functions for salsa Db to impl EffectInfo.
This required quite a bit of refactoring. The largest being moving Nst
to name res and renaming it. It now has access to types from name res,
which makes sense it is responsible for exposing those mappings to the
downstream code. But that means down stream code has to depend on
`aiahr-analysis` and not just `aiahr-core`.