Create salsa API for tc.
At the moment this is just one function `type_scheme_of`. With the
expectation being that lower-ir will call it lazily as it needs the
types of terms/items it encounters.

If the need arises it's easy to create a `type_check_module` that just
calls `type_scheme_of` on every item it contains.