Create salsa API for lower-ir crate.
Refactor traits in lower-ir to be able to impl them for salsa Db. This
is mostly removing lifetimes and adding ModuleId/ItemId to parameters.

It also marks the `Db` parameter in `LowerCtx` as `?Sized` allowing `Db`
to be `dyn crate::Db` directly. Before this it had to be `&dyn
crate::Db` so it was sized and this reduces trait impls we can have.