Migrate parser to use `cst::indexed` instead of `cst` types.
This turns out to be a pretty substantial change. Since the arena
allocator worked by ref we could thread it through as a module. However
that doesn't work for `&mut CstIndxAlloc` as it's has to be used
exclusively.

Instead we now parse into a poor man's State monad. Instead of a
`Vec<Item>` we parse a `StateM<CstIdxAlloc, Vec<Item>>` which we then
pass a `&mut CstIdxAlloc` to produce a `Vec<Item>`. This is just a
temporary solution until chumsky 1.0.0 is released with support for
threading a State param through the `Parser` itself.