Replace `ModuleId` with `Module`.
This is a large step in integrating our compiler with salsa. Previously
we held a ModuleId and would use it to lookup `Module` struct to access
the various salsa structs and methods we need.

Now nameres embeds a `Module` directly in the Id tuples it produces.
This allows downstream crates to access their needed salsa structs more
directly.

`Module` and `ModuleId` are both represented as a wrapper around an
integer Id, so we don't need to worry about performance hits from
embedding a different sized Module into id tuples.