Introduce identifier that encompass their ID and scope.
Introduce TermName, EffectName, and EffectOpName strucs that contain the
syntax name of their construct and the enclosing scope (Module or
Effect). These are interned within salsa so they can be used as Ids for salsa
tracked methods.

We now produce these structs instead of their counterparts (ItemId,
EffectId, EffectOpId) in nameres. This is a large cross cutting change
but relatively minor in each crate we change.

Because of these changes to nameres we had to reconfigure some tests.