Implement scoped rows.
This is a large change to the type checker context.

This commit is a bit of mid way point as we still need to do some work
to implement scoped rows. However this is already a giant change and
this was a natural break point to split up the commits.