Start introducing scoped rows into lower-ir.
Reorder `Row` generics and remove default for `Sema` parameter.

In trying to introduce scoped row evidence to lower-ir it becomes clear
to me that we need to be more structured in how we deal with evidence.
I'm putting a hold on this change in a partial state, to take a pass
over type inference and pass more metadata around evidence from type
inference to lowering. This should make it easier to handle scoped rows
than it is currently.