Output required evidence and wrappers from type checking.
We not track evidence and `wrappers` during constraint generation,
similar to how we track variables to types and terms to types. These
will be used to help construct lowered terms in a more principaled way.

Previously in lowering we had to infer row evidence usage on the fly and
determine if it was solved, and then if we had already solved it or not.
This was workable but will lead to long term issues. For example if we
want to say, have one module with all our solved row evidence (or some
pre-defined set of builtin row evidence). This approach won't work
because we locally let bind solved row evidence.

Now that we collect all required evidence up front, we can emit a
parameter for each needed evidence during lowering. If the parameter is
for solved evidence we can produce the solved evidence as needed and
inline it into the lowered body (same as we'd do if it was locally let
bound). This will also to less noisy IR which is a nice bonus.