Create `*_rowlikes` operations for closed rows.
Previously Simple/Scoped closed rows shared code that worked on
`ClosedRow`. This code has been further generalized to work on anything
that is "row-like", in that it is a `&[RowLabel]` and a `&[V]` for some
`V` that isn't necessairly a type.

This will be useful for generating row evidence at the reducir level. We
can use our rowlike operators to calculate where indices should go when
generating our concat/branch/prj/inj combinators. This avoids having to
reimplement what ends up being the same logic to combine rows of types
and trivial rows.