Initial implementation of type_check for ReducIr.
Because ReducIr is explicitly typed, type check is literally just
checking that the present types are consistent with each other. It does
no inference. This is relatively cheap, especially compared to
inference.

Type checking revealed a whole host of type errors in our generated row
evidence that were bugs. Types work! This commmit fixes a lot of those
as well.