Mark `Case` nodes with their return type.
This lets us remove the NeverTy. Type equality can now be done
syntactically again.

Eventually we can also use this as a cache so we don't have to type check every branch we can just trust the
type in the constructor. We don't do this yet though as we'd have to
check our invariants on construction first.