Typechecking effect and row evidence works now.
This required a lot of changes. Typechecking now tracks an effect row as
a set of possible effects, instead of a set of performed effects. This
means values infer an open effect row and we equate effect rows when
they meet instead of combining them.

Alongside this change we now track return type in the effect row.

Lowering handlers into ReducIr had to change. We now always bind a
return variable for effect sigs at the AST level and use this to
construct the correct type for handlers at the IR level by instantiating
that variable.

Extensive changes to how we lower effect code. Most of these are just
getting the types to line up but we have to pay special attention to
handlers and return type of effects. We use these to construct the
struct that gets inserted into the effect row and the type of the
continuation that is captured by yielding.

Minor bugfixes to row evidence typing and instantiating row evidence at
the correct types.