Simplify ReducIr terms via naive inlining.
This is an extremely conservative inlining procedure. It will only
inline know function applications if their argument is a value. It does
not make use of any occurrence info to determine if inlining is
profitable or not and does not check size of inlined term for possible
code explosion.

Expectation is we'll have to replace this with something more
sophisticated, but this gets us off the ground and makes terms easier to
work with down stream by making obvious optimizations.