Initial prototype of emittting wasm.
We produce a wasm module out of an optimized reducir module.
That wasm module does not work yet. Have to figure out some issues
around closures and generating calls to functions that don't exist.

However I wanted to put a pin in this work to go and work on inlining
the monadic combinators so we don't have to use wasm implicits for them.
We might end up doing that anyways for effiency but would prefer to
avoid it if possible.