Add support for entry points.
Entry points have a known type, and are checked against that type
instead of inferred. This type is an Int with no effects. Because it has
no effects it generates nicer code in the wasm.

Add a wasm binary that executes a file using wastime. This is intended
for ad hoc testing.