Fixed lowering singleton product literal.
When we lower a singleton product it produces a label with an uncast
RowTy. This is the only case when a RowTy can be produced so we assume
this is the case and cast it in lower-reducir.

This is hack that we will surely pay for down the line. My apologies.