Use evidence to select operation out of effect handler.
Previously we used a hardcoded index to select op out of effect handler
in lowering. This presents an issue for handlers with a single
operation. Because they are a singleton row they get unwrapped, but once
our indexing logic doesn't recognize that and tries to index into the
unwrapped row.

To remedy this, we generate a piece of evidence during type checking
that selects the handler operation out of the effect handler. Because
this is just a normal data evidence it handles any unwrapping required
for us. Then during lowering effects we use the generated evidence to
select our handler instead of an index.