Fix wasm emitted for match exprs.
When emitting a match only the topmost block should be typed. The
interior blocks should all have the empty type. This is so that the
`br_table` ins that targets them all produces an empty type.

This solves the issue we had previously where `br_table` required having
an extra i32 on the stack when it was called. Now br_table returns the
empty type correctly and whatever case we jump into branches to the
topmost block and produces a value based on it's type.