Treat evv as a normal variable during lowering.
Evv was the only variable we re-used during lowering (and monadic
lowering). This proved to be problematic in the simplifier as it causes
headaches for inlining. Even worse it doesn't actually do anything for
us (aside from being simple at the time). Remove that and instead use
normal variables for evv that are rebound as we change evv.